requirements.txt
Flask==3.0.0
Pillow==10.1.0
rembg[cpu,cli]==2.0.58
click==8.1.7
filetype==1.2.0
flask-limiter==3.5.0
